Use dynamic imports to reduce initial load times
Remove videojs
