+++ web/src/App.tsx
@@ -2,20 +2,23 @@ import Providers from "@/context/providers";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Wrapper from "@/components/Wrapper";
 import Sidebar from "@/components/navigation/Sidebar";
-import Live from "@/pages/Live";
-import Export from "@/pages/Export";
-import Storage from "@/pages/Storage";
-import System from "@/pages/System";
-import ConfigEditor from "@/pages/ConfigEditor";
-import Logs from "@/pages/Logs";
-import NoMatch from "@/pages/NoMatch";
-import Settings from "@/pages/Settings";
-import UIPlayground from "./pages/UIPlayground";
-import Events from "./pages/Events";
+
 import { isDesktop, isMobile } from "react-device-detect";
 import Statusbar from "./components/Statusbar";
 import Bottombar from "./components/navigation/Bottombar";
-import SubmitPlus from "./pages/SubmitPlus";
+import { Suspense, lazy } from "react";
+
+const Live = lazy(() => import("@/pages/Live"));
+const Events = lazy(() => import("@/pages/Events"));
+const Export = lazy(() => import("@/pages/Export"));
+const Storage = lazy(() => import("@/pages/Storage"));
+const SubmitPlus = lazy(() => import("@/pages/SubmitPlus"));
+const ConfigEditor = lazy(() => import("@/pages/ConfigEditor"));
+const System = lazy(() => import("@/pages/System"));
+const Settings = lazy(() => import("@/pages/Settings"));
+const UIPlayground = lazy(() => import("@/pages/UIPlayground"));
+const Logs = lazy(() => import("@/pages/Logs"));
+const NoMatch = lazy(() => import("@/pages/NoMatch"));
 
 function App() {
   return (
@@ -30,19 +33,21 @@ function App() {
               id="pageRoot"
               className={`absolute top-2 right-0 overflow-hidden ${isMobile ? "left-0 bottom-16" : "left-16 bottom-8"}`}
             >
-              <Routes>
-                <Route path="/" element={<Live />} />
-                <Route path="/events" element={<Events />} />
-                <Route path="/export" element={<Export />} />
-                <Route path="/storage" element={<Storage />} />
-                <Route path="/plus" element={<SubmitPlus />} />
-                <Route path="/system" element={<System />} />
-                <Route path="/settings" element={<Settings />} />
-                <Route path="/config" element={<ConfigEditor />} />
-                <Route path="/logs" element={<Logs />} />
-                <Route path="/playground" element={<UIPlayground />} />
-                <Route path="*" element={<NoMatch />} />
-              </Routes>
+              <Suspense>
+                <Routes>
+                  <Route path="/" element={<Live />} />
+                  <Route path="/events" element={<Events />} />
+                  <Route path="/export" element={<Export />} />
+                  <Route path="/storage" element={<Storage />} />
+                  <Route path="/plus" element={<SubmitPlus />} />
+                  <Route path="/system" element={<System />} />
+                  <Route path="/settings" element={<Settings />} />
+                  <Route path="/config" element={<ConfigEditor />} />
+                  <Route path="/logs" element={<Logs />} />
+                  <Route path="/playground" element={<UIPlayground />} />
+                  <Route path="*" element={<NoMatch />} />
+                </Routes>
+              </Suspense>
             </div>
           </div>
         </Wrapper>
